[WASH-981] Tozny Flutter plugin release 0.0.5
- Updates the Android plugin's build.gradle to use the recently
  released version 7.2.3
- Updates the iOS plugin to return the correct Unix timestamp
  beginning from 1 January 1970
- Updates tozny_model to expect an Int for the expiry field
  of an AgentToken

Note that the iOS plugin require two levels of indirection
in order to respect both Swift's Codable protocol and what
Tozny's models expect. This commit introduces an intermediate
Codable class for the E3db.AgentToken returned by the Swift SDK,
and a FlutterAgentToken class that is initialized using the
intermediate class.

+++ lib/tozny_model.dart
@@ -219,7 +219,7 @@ class AgentToken {
   @JsonKey(name: "token_type")
   String tokenType;
   @JsonKey(name: "expiry")
-  double expiry;
+  int expiry;
 
   AgentToken(
     this.accessToken,
